feat(#108): Style EoRepresentationTest
