+++ src/main/java/org/eolang/jeo/representation/asm/XmlBytecode.java
@@ -1,3 +1,26 @@
+/*
+ * The MIT License (MIT)
+ *
+ * Copyright (c) 2016-2023 Objectionary.com
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included
+ * in all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NON-INFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ */
 package org.eolang.jeo.representation.asm;
 
 import com.jcabi.log.Logger;
@@ -11,10 +34,21 @@ import org.w3c.dom.NamedNodeMap;
 import org.w3c.dom.Node;
 import org.w3c.dom.NodeList;
 
-public class XmlBytecode extends ClassWriter {
+/**
+ * XML to Java bytecode.
+ * @since 0.1.0
+ */
+public final class XmlBytecode extends ClassWriter {
 
+    /**
+     * XML.
+     */
     private final XML xml;
 
+    /**
+     * Constructor.
+     * @param xml XML.
+     */
     public XmlBytecode(final XML xml) {
         super(ClassWriter.COMPUTE_MAXS);
         this.xml = xml;
@@ -32,7 +66,7 @@ public class XmlBytecode extends ClassWriter {
             final String content = name.getTextContent();
             final String[] split = content.split("__");
             this.visit(
-                new AsmVersion().java(),
+                new DefaultVersion().java(),
                 Integer.parseInt(split[0]),
                 String.valueOf(split[1]),
                 null,
@@ -174,5 +208,4 @@ public class XmlBytecode extends ClassWriter {
             return base.getNodeValue().equals("seq");
         }
     }
-
 }
